Combine edit and create stream pages
They act the same way, and with this update we can effortlessly take the user to the newly created stream's page as soon as it's available (coming up!).
